IOSTests/requirements.txt
Appium-Python-Client==3.1.1
pytest==8.0.0
pytest-html==4.1.1
allure-pytest==2.13.2
pyotp==2.9.0